+++ AddOns/TeleportMenu/CHANGELOG.md
@@ -1,7 +1,15 @@
 # Teleport Menu
 
-## [v10.3](https://github.com/Justw8/TeleportMenu/tree/v10.3) (2025-02-27)
-[Full Changelog](https://github.com/Justw8/TeleportMenu/compare/v10.2...v10.3) [Previous Releases](https://github.com/Justw8/TeleportMenu/releases)
+## [v10.4](https://github.com/Justw8/TeleportMenu/tree/v10.4) (2025-04-08)
+[Full Changelog](https://github.com/Justw8/TeleportMenu/compare/v10.3...v10.4) [Previous Releases](https://github.com/Justw8/TeleportMenu/releases)
 
-- Fix: Wormhole teleport error when checking usableWormholes length (#94)  
-- Add Kirin Tor rings (#92)  
+- Update vscode setting files  
+- Add Ring-Bound Hourglass  
+- Add Fractured Necrolyte Skull  
+- Add Gallagio Loyalty Rewards Cards  
+- Add Boots of the Bay  
+- Add luacheck and code updates  
+- Update Items/zhCN/zhTW (#108)  
+- Set the season to be static to fix loading seasonal teleports (#106)  
+- Add Argent Crusader's Tabard (#104)  
+- Add Explosive Hearthstone (#96)